+++ server/queries/character.js
@@ -1,0 +1,13 @@
+const characterData = `
+    query {
+      characterData {
+        character(name: "Marbin", serverSlug: "frostmourne", serverRegion: "us") {
+          id
+          name
+          level
+        }
+      }
+    }
+  `
+
+export { characterData }
